backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
reportlab==4.1.0
alembic==1.13.1
pydantic[email]==2.10.6
python-dotenv==1.0.1
email-validator==2.2.0
python-dateutil
